+++ src/oxl_ansible_webui/aw/api_endpoints/credentials.py
@@ -1,3 +1,5 @@
+from random import choice
+
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.utils import IntegrityError
 from rest_framework.views import APIView
@@ -11,8 +13,11 @@ from aw.model.permission import CHOICE_PERMISSION_READ, CHOICE_PERMISSION_WRITE,
 from aw.api_endpoints.base import API_PERMISSION, get_api_user, GenericResponse, BaseResponse, api_docs_delete, \
     api_docs_put, api_docs_post, validate_no_xss, GenericErrorResponse, response_data_if_changed, API_PARAM_HASH
 from aw.utils.permission import has_credentials_permission, has_manager_privileges
+from aw.utils.util import is_null, overwrite_and_delete_file, write_file_0600
+from aw.execute.play_credentials import get_pwd_file
 from aw.config.hardcoded import SECRET_HIDDEN
-from aw.utils.util import is_null
+from aw.execute.util import get_path_run
+from aw.utils.subps import process
 from aw.base import USERS
 
 
@@ -564,5 +569,92 @@ class APIJobUserCredentialsItem(APIView):
 
         return Response(
             data={'msg': f"User-credentials '{credentials.name}' updated", 'id': credentials_id},
+            status=200,
+        )
+
+
+class VaultEncryptRequest(BaseResponse):
+    plaintext = serializers.CharField(required=True)
+    credentials_id = serializers.IntegerField(required=True)
+    credentials_type = serializers.ChoiceField(required=True, choices=['shared', 'user'])
+
+
+class VaultEncryptResponse(GenericResponse):
+    ciphertext = serializers.CharField()
+
+
+class APIVaultEncrypt(APIView):
+    http_method_names = ['post']
+    serializer_class = VaultEncryptResponse
+    permission_classes = API_PERMISSION
+
+    @extend_schema(
+        request=VaultEncryptRequest,
+        responses={
+            200: OpenApiResponse(JobUserCredentialsReadResponse, description='Successfully encrypted data'),
+            404: OpenApiResponse(GenericErrorResponse, description='Credentials not exist or have no vault-secret set'),
+            500: OpenApiResponse(GenericErrorResponse, description='Failed to encrypt data'),
+        },
+        summary='Encrypt a secret using Ansible-Vault.',
+        operation_id='credentials_vault_encrypt',
+    )
+    def post(self, request):
+        user = get_api_user(request)
+
+        serializer = VaultEncryptRequest(data=request.data)
+        if not serializer.is_valid():
+            return Response(data={'error': 'Invalid vault-encrypt request'}, status=400)
+
+        credentials_id = serializer.validated_data['credentials_id']
+        if serializer.validated_data['credentials_type'] == 'user':
+            credentials = _get_user_creds(credentials_id, user)
+
+        else:
+            credentials = _get_shared_creds(credentials_id)
+
+        if credentials is None:
+            return Response(
+                data={'error': f"Credentials with ID {credentials_id} do not exist or are inaccessible"},
+                status=404,
+            )
+
+        print("FILE", credentials.vault_file, "PASS", credentials.vault_pass)
+        if is_null(credentials.vault_file) and is_null(credentials.vault_pass):
+            return Response(
+                data={'error': f"Credentials with ID {credentials_id} have no vault-secret configured"},
+                status=404,
+            )
+
+        vault_file = credentials.vault_file
+        if is_null(vault_file):
+            path_run = get_path_run()
+            if not path_run.is_dir():
+                path_run.mkdir()
+
+            vault_file = get_pwd_file(path_run=path_run, attr='vault_pass')
+            write_file_0600(
+                file=vault_file,
+                content=getattr(credentials, 'vault_pass'),
+            )
+
+        cmd = ['ansible-vault', 'encrypt_string', f'--vault-password-file={vault_file}']
+        if not is_null(credentials.vault_id):
+            cmd.append(f'--vault-id={credentials.vault_id}')
+
+        cmd.append(serializer.validated_data['plaintext'])
+
+        result = process(cmd=cmd, timeout_sec=2)
+
+        if result['rc'] != 0:
+            return Response(
+                data={'error': "Failed to Ansible-Vault-encrypt data"},
+                status=500,
+            )
+
+        return Response(
+            data={
+                'msg': 'Successfully Ansible-Vault-encrypted data',
+                'ciphertext': result['stdout'],
+            },
             status=200,
         )