+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Fixes:
 
 * Git-Origin not accepting long URLs (100 => 500 characters)
 * Verbosity missing from Execution-Dialogue [#47](https://github.com/O-X-L/ansible-webui/issues/47)
+* Fix Job-Edit File-Browsing caching repository-content
 
 Features: