+++ CHANGELOG.md
@@ -8,6 +8,12 @@ tbd
 
 ## Version 0
 
+### 0.9.0
+
+Features:
+
+* API and Form to Ansible-Vault-Encrypt data [#7](https://github.com/O-X-L/ansible-webui/issues/7)
+
 ### 0.8.9
 
 Fixes: